made deactivation/deletion of the last superuser disabled

+++ moztrap/model/core/auth.py
@@ -38,6 +38,9 @@ class User(BaseUser):
         https://code.djangoproject.com/ticket/16128.
 
         """
+        if (self.is_superuser == True and self.is_active == True and
+            User.objects.filter(is_superuser=True).filter(is_active=True).count() == 1):
+            return
         # @@@ Django ticket 16128, hopefully fixed in 1.4?
         # RegistrationProfile's FK is to Django's user model, not ours
         RegistrationProfile.objects.filter(user=self).delete()
@@ -54,6 +57,19 @@ class User(BaseUser):
         """Deactivate this user."""
         self.is_active = False
         self.save(force_update=True)
+
+
+    def save(self, force_insert=False, force_update=False, using=None):
+        if (not force_insert and self.is_superuser and not self.is_active and
+            User.objects.filter(is_superuser=True).filter(is_active=True).count() == 1):
+            from django.shortcuts import get_object_or_404
+            user = get_object_or_404(User, pk=self.id)
+            # check whether the user is exactly the last `active` superuser or not
+            if user.is_active:
+                self.is_active = True
+
+        super(User, self).save(force_insert=force_insert,
+            force_update=force_update, using=using)
 
 
     @property
